// pokemons/src/app/interfaces/pokemon.ts
export interface PaginationData {
    itemsPerPage: number,
    currentPage: number,
    totalItems: number
}

export interface ItemPokemon {
    name: string,
    url: string,
    image: string,
    favourite?: boolean
}

export interface PokemonModel {
    abilities: Abilities[];
    base_experience: number;
    forms: Form[];
    game_indices: GameIndice[];
    height: number;
    held_items: HeldItem[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Move[];
    name: string;
    order: number;
    past_types: any[];
    species: Species;
    sprites: Sprites;
    stats: Stat[];
    types: Type[];
    weight: number;
    favourite?:boolean;
}
    
    
    export interface Ability {
        name: string;
        url: string;
    }

    export interface Abilities {
        ability: Ability;
        is_hidden: boolean;
        slot: number;
    }

    export interface Form {
        name: string;
        url: string;
    }

    export interface Version {
        name: string;
        url: string;
    }

    export interface GameIndice {
        game_index: number;
        version: Version;
    }

    export interface Item {
        name: string;
        url: string;
    }

    export interface Version2 {
        name: string;
        url: string;
    }

    export interface VersionDetail {
        rarity: number;
        version: Version2;
    }

    export interface HeldItem {
        item: Item;
        version_details: VersionDetail[];
    }

    export interface Move2 {
        name: string;
        url: string;
    }

    export interface MoveLearnMethod {
        name: string;
        url: string;
    }

    export interface VersionGroup {
        name: string;
        url: string;
    }

    export interface VersionGroupDetail {
        level_learned_at: number;
        move_learn_method: MoveLearnMethod;
        version_group: VersionGroup;
    }

    export interface Move {
        move: Move2;
        version_group_details: VersionGroupDetail[];
    }

    export interface Species {
        name: string;
        url: string;
    }

    export interface DreamWorld {
        front_default: string;
        front_female?: any;
    }

    export interface Home {
        front_default: string;
        front_female: string;
        front_shiny: string;
        front_shiny_female: string;
    }

    export interface OfficialArtwork {
        front_default: string;
    }

    export interface Other {
        dream_world: DreamWorld;
        home: Home;
        // official-artwork: OfficialArtwork;
    }
// ...
    export interface Versions {
        // generation-i: GenerationI;
        // generation-ii: GenerationIi;
        // generation-iii: GenerationIii;
        // generation-iv: GenerationIv;
        // generation-v: GenerationV;
        // generation-vi: GenerationVi;
        // generation-vii: GenerationVii;
        // generation-viii: GenerationViii;
    }

    export interface Sprites {
        back_default: string;
        back_female: string;
        back_shiny: string;
        back_shiny_female: string;
        front_default: string;
        front_female: string;
        front_shiny: string;
        front_shiny_female: string;
        other: Other;
        versions: Versions;
    }

    export interface Stat2 {
        name: string;
        url: string;
    }

    export interface Stat {
        base_stat: number;
        effort: number;
        stat: Stat2;
    }

    export interface Type2 {
        name: string;
        url: string;
    }

    export interface Type {
        slot: number;
        type: Type2;
    }